Safari-609.1.11.1/LayoutTests/fast/loader/font-face-empty.html

<style>
    @font-face {
        font-family: TestFont;
        src: url(data:application/x-truetype-font,) format(truetype);
    }
</style>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();
</script>
<p>This is a test for <a href="http://bugs.webkit.org/show_bug.cgi?id=16996">Bug 16996: Crash in
createFontCustomPlatformData when loading 0-byte font via @font-face</a>.</p>
<p style="font-family: TestFont">`1234567890-=qwertyuiop[]\asdfghjkl;'zxcvbnm,./~!@#$%^&amp;*()_+QWERYUIOP{}|ASDFGHJKL:"ZXCVBNM&lt;&gt;?</p>
